LUKS1 tests: generating test-containers on-the-fly.

+++ tools/test.sh
@@ -22,16 +22,20 @@ VC_MODES="13711 13712 13713 13721 13722 13723 13731 13732 13733 13741 13742 1374
 NOCHECK_ENCODING="16800 22000"
 
 # List of LUKS modes which have test containers
-LUKS_MODES="14600 29511 29512 29513 29521 29522 29523 29531 29532 29533 29541 29542 29543"
+LUKS1_LEGACY_MODE="14600"
+LUKS1_MODES="29511 29512 29513 29521 29522 29523 29531 29532 29533 29541 29542 29543"
+LUKS1_ALL_MODES="${LUKS1_LEGACY_MODE} ${LUKS1_MODES}"
 
 # List of LUKS2 modes which have test containers
 LUKS2_MODES="34100"
+
+LUKS_MODES="${LUKS1_MODES} ${LUKS2_MODES}"
 
 # Cryptoloop mode which have test containers
 CL_MODES="14511 14512 14513 14521 14522 14523 14531 14532 14533 14541 14542 14543 14551 14552 14553"
 
 HASH_TYPES=$(ls "${TDIR}"/test_modules/*.pm | sed -E 's/.*m0*([0-9]+).pm/\1/')
-HASH_TYPES="${HASH_TYPES} ${TC_MODES} ${VC_MODES} ${LUKS_MODES} ${LUKS2_MODES} ${CL_MODES}"
+HASH_TYPES="${HASH_TYPES} ${TC_MODES} ${VC_MODES} ${LUKS1_ALL_MODES} ${LUKS2_MODES} ${CL_MODES}"
 HASH_TYPES=$(echo -n "${HASH_TYPES}" | tr ' ' '\n' | sort -u -n | tr '\n' ' ')
 
 VECTOR_WIDTHS="1 2 4 8 16"
@@ -165,6 +169,24 @@ function is_in_array()
   return 1
 }
 
+function has_multi_hash()
+{
+  # no multi hash checks for these modes (because we only have 1 hash for each of them)
+  if   [ "${hash_type}" -eq 14000 ]; then
+    return 0;
+  elif [ "${hash_type}" -eq 14100 ]; then
+    return 0;
+  elif [ "${hash_type}" -eq 14600 ]; then
+    return 0;
+  elif [ "${hash_type}" -eq 14900 ]; then
+    return 0;
+  elif [ "${hash_type}" -eq 15400 ]; then
+    return 0;
+  fi
+
+  # all other modes have multi hash
+  return 1;
+}
 function init()
 {
   if [ "${PACKAGE}" -eq 1 ]; then
@@ -186,7 +208,7 @@ function init()
     return 0
   fi
 
-  if is_in_array "$hash_type" ${LUKS_MODES} || { is_in_array "$hash_type" ${LUKS2_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] }; then
+  if { is_in_array "$hash_type" ${LUKS1_ALL_MODES} || is_in_array "$hash_type" ${LUKS2_MODES}; } && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ; then
     which 7z &>/dev/null
     if [ $? -eq 1 ]; then
       echo "ATTENTION: 7z is missing. Skipping download and extract luks test files."
@@ -195,7 +217,7 @@ function init()
   fi
 
   #LUKS1
-  if is_in_array "$hash_type" ${LUKS_MODES}; then
+  if is_in_array "$hash_type" ${LUKS1_ALL_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ; then
     luks_tests_folder="${TDIR}/luks_tests/"
 
     if [ ! -d "${luks_tests_folder}" ]; then
@@ -324,7 +346,7 @@ function init()
   sed 's/^echo *|.*$//'       "${cmd_file}" | awk '{print $2}'                                                                    > "${OUTD}/${hash_type}_passwords.txt"
   sed 's/^echo *|/echo "" |/' "${cmd_file}" | awk '{t="";for(i=10;i<=NF;i++){if(t){t=t" "$i}else{t=$i}};print t}' | cut -d"'" -f2 > "${OUTD}/${hash_type}_hashes.txt"
 
-  if [ "${hash_type}" -eq 34100 ]; then
+  if is_in_array "${hash_type}" ${LUKS_MODES}; then
     # 34100 LUKS2 dynamically generates filenames, we need to cat those to get the hashes
     mv "${OUTD}/${hash_type}_hashes.txt" "${OUTD}/${hash_type}_hashes.tmp"
     cat "${OUTD}/${hash_type}_hashes.tmp" | while read f; do cat $f; done > "${OUTD}/${hash_type}_hashes.txt"
@@ -437,6 +459,9 @@ function init()
   # generate multiple pass/hash foreach len (2 to 8)
   if [ "${MODE}" -ge 1 ]; then
 
+    # no multi hash checks for these modes (because we only have 1 hash for each of them)
+    ! has_multi_hash || return
+
     i=2
 
     while [ "$i" -lt 9 ]; do
@@ -460,7 +485,7 @@ function init()
       sed 's/^echo *|.*$//'       "${cmd_file}" | awk '{print $2}'                                                                    > "${OUTD}/${hash_type}_passwords_multi_${i}.txt"
       sed 's/^echo *|/echo "" |/' "${cmd_file}" | awk '{t="";for(i=10;i<=NF;i++){if(t){t=t" "$i}else{t=$i}};print t}' | cut -d"'" -f2 > "${OUTD}/${hash_type}_hashes_multi_${i}.txt"
 
-      if [ "${hash_type}" -eq 34100 ]; then
+      if is_in_array "${hash_type}" ${LUKS_MODES}; then
         mv "${OUTD}/${hash_type}_hashes_multi_${i}.txt" "${OUTD}/${hash_type}_hashes_multi_${i}.tmp"
         cat  "${OUTD}/${hash_type}_hashes_multi_${i}.tmp" | while read f; do cat $f; done > "${OUTD}/${hash_type}_hashes_multi_${i}.txt"
         rm  "${OUTD}/${hash_type}_hashes_multi_${i}.tmp"
@@ -638,7 +663,7 @@ function attack_0()
         break
       fi
 
-      if [ "${file_only}" -eq 1 ]; then
+      if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
         temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
 
@@ -671,7 +696,7 @@ function attack_0()
       echo "${output}" >> "${OUTD}/logfull.txt"
 
       if [ "${ret}" -eq 0 ]; then
-        if   [ "${hash_type}" -ne  34100 ]; then
+        if ! (is_in_array "${hash_type}" ${LUKS_MODES}); then
           if [ "${pass_only}" -eq 1 ]; then
             search=":${pass}"
           else
@@ -682,10 +707,10 @@ function attack_0()
           newRet=$?
         fi
 
-        if [[ "${newRet}" -eq 2 || "${hash_type}" -eq 34100 ]]; then
+        if [[ "${newRet}" -eq 2 ]]; then
           # out-of-memory, workaround
 
-          if [ "${hash_type}" -eq 34100 ]; then
+          if is_in_array "${hash_type}" ${LUKS_MODES}; then
             search="$(cat ${hash} | tr -d '\n'):${pass}" #hash is a filename for 34100
             echo "${output}" | grep -E '^\$luks\$' | head -1 > tmp_file_out #cracked hash from hashcat output
           else
@@ -713,7 +738,7 @@ function attack_0()
           fi
         fi
 
-        if [ "${newRet}" -ne 0 ]; then
+        if [[ "${newRet}" -ne 0 ]]; then
           if [ "${newRet}" -eq 2 ]; then
             ret=20
           else
@@ -735,7 +760,7 @@ function attack_0()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -746,6 +771,9 @@ function attack_0()
 
   # multihash
   if [ "${MODE}" -ne 0 ]; then
+
+    # no multi hash checks for these modes (because we only have 1 hash for each of them)
+    ! has_multi_hash || return
 
     e_ce=0
     e_rs=0
@@ -760,7 +788,7 @@ function attack_0()
 
     # if file_only -> decode all base64 "hashes" and put them in the temporary file
 
-    if [ "${file_only}" -eq 1 ]; then
+    if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
       temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
       rm -f "${temp_file}"
@@ -795,7 +823,7 @@ function attack_0()
 
         pass=$(sed -n ${i}p "${OUTD}/${hash_type}_passwords.txt")
 
-        if   [ "${hash_type}" -ne  34100 ]; then
+        if is_in_array "${hash_type}" ${LUKS_MODES}; then
           if [ "${pass_only}" -eq 1 ]; then
             search=":${pass}"
           else
@@ -807,7 +835,7 @@ function attack_0()
           newRet=$?
         fi
 
-        if [ "${hash_type}" -eq 34100 ]; then
+        if is_in_array "${hash_type}" ${LUKS_MODES}; then
           search="$(echo ${hash} | tr -d '\n'):${pass}" #hash is read from file already
           to_search_in_out="$(echo ${hash} | head -c 200)"
           echo "${output}" | grep -F "$to_search_in_out" | head -1 > tmp_file_out #cracked hash from hashcat output
@@ -856,7 +884,7 @@ function attack_0()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -910,7 +938,7 @@ function attack_1()
 
       if [ $i -gt ${min} ]; then
 
-        if [ "${file_only}" -eq 1 ]; then
+        if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
           temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
 
@@ -1050,7 +1078,7 @@ function attack_1()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -1063,16 +1091,7 @@ function attack_1()
   if [ "${MODE}" -ne 0 ]; then
 
     # no multi hash checks for these modes (because we only have 1 hash for each of them)
-
-    if   [ "${hash_type}" -eq 14000 ]; then
-      return
-    elif [ "${hash_type}" -eq 14100 ]; then
-      return
-    elif [ "${hash_type}" -eq 14900 ]; then
-      return
-    elif [ "${hash_type}" -eq 15400 ]; then
-      return
-    fi
+    ! has_multi_hash || return
 
     e_ce=0
     e_rs=0
@@ -1095,7 +1114,7 @@ function attack_1()
 
     # if file_only -> decode all base64 "hashes" and put them in the temporary file
 
-    if [ "${file_only}" -eq 1 ]; then
+    if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
       temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
       rm -f "${temp_file}"
@@ -1172,7 +1191,7 @@ function attack_1()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -1232,7 +1251,7 @@ function attack_3()
         fi
       fi
 
-      if [ "${file_only}" -eq 1 ]; then
+      if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
         temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
 
@@ -1304,7 +1323,7 @@ function attack_3()
         line_dict=$(sed -n ${i}p "${dict}")
 
 
-        if   [ "${hash_type}" -eq  34100 ]; then
+        if is_in_array "${hash_type}" ${LUKS_MODES}; then
           if [ "${pass_only}" -eq 1 ]; then
             search=":${line_dict}"
           else
@@ -1363,7 +1382,7 @@ function attack_3()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -1376,16 +1395,7 @@ function attack_3()
   if [ "${MODE}" -ne 0 ]; then
 
     # no multi hash checks for these modes (because we only have 1 hash for each of them)
-
-    if   [ "${hash_type}" -eq 14000 ]; then
-      return
-    elif [ "${hash_type}" -eq 14100 ]; then
-      return
-    elif [ "${hash_type}" -eq 14900 ]; then
-      return
-    elif [ "${hash_type}" -eq 15400 ]; then
-      return
-    fi
+    ! has_multi_hash || return
 
     e_ce=0
     e_rs=0
@@ -1419,7 +1429,7 @@ function attack_3()
 
     # if file_only -> decode all base64 "hashes" and put them in the temporary file
 
-    if [ "${file_only}" -eq 1 ]; then
+    if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
       temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
       rm -f "${temp_file}"
@@ -1786,7 +1796,7 @@ function attack_3()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -1884,7 +1894,7 @@ function attack_6()
 
       if [ ${i} -gt ${min} ]; then
 
-        if [ "${file_only}" -eq 1 ]; then
+        if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
           temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
 
@@ -2027,7 +2037,7 @@ function attack_6()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -2043,16 +2053,7 @@ function attack_6()
   if [ "${MODE}" -ne 0 ]; then
 
     # no multi hash checks for these modes (because we only have 1 hash for each of them)
-
-    if   [ "${hash_type}" -eq 14000 ]; then
-      return
-    elif [ "${hash_type}" -eq 14100 ]; then
-      return
-    elif [ "${hash_type}" -eq 14900 ]; then
-      return
-    elif [ "${hash_type}" -eq 15400 ]; then
-      return
-    fi
+    ! has_multi_hash || return
 
     e_ce=0
     e_rs=0
@@ -2103,7 +2104,7 @@ function attack_6()
 
       # if file_only -> decode all base64 "hashes" and put them in the temporary file
 
-      if [ "${file_only}" -eq 1 ]; then
+      if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
         temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
         rm -f "${temp_file}"
@@ -2179,7 +2180,7 @@ function attack_6()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -2274,7 +2275,7 @@ function attack_7()
 
       if [ ${i} -gt ${min} ]; then
 
-        if [ "${file_only}" -eq 1 ]; then
+        if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
           temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
 
@@ -2470,7 +2471,7 @@ function attack_7()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -2486,16 +2487,7 @@ function attack_7()
   if [ "${MODE}" -ne 0 ]; then
 
     # no multi hash checks for these modes (because we only have 1 hash for each of them)
-
-    if   [ "${hash_type}" -eq 14000 ]; then
-      return
-    elif [ "${hash_type}" -eq 14100 ]; then
-      return
-    elif [ "${hash_type}" -eq 14900 ]; then
-      return
-    elif [ "${hash_type}" -eq 15400 ]; then
-      return
-    fi
+    ! has_multi_hash || return
 
     e_ce=0
     e_rs=0
@@ -2557,7 +2549,7 @@ function attack_7()
 
       # if file_only -> decode all base64 "hashes" and put them in the temporary file
 
-      if [ "${file_only}" -eq 1 ]; then
+      if [ "${file_only}" -eq 1 ] && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
 
         temp_file="${OUTD}/${hash_type}_filebased_only_temp.txt"
         rm -f "${temp_file}"
@@ -2657,7 +2649,7 @@ function attack_7()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -2840,7 +2832,7 @@ function cryptoloop_test()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -3218,7 +3210,7 @@ function truecrypt_test()
       msg="Compare Error"
     elif [ "${e_rs}" -ne 0 ]; then
       msg="Skip"
-    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+    elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
       msg="Error"
     elif [ "${e_to}" -ne 0 ]; then
       msg="Warning"
@@ -3319,7 +3311,7 @@ function veracrypt_test()
     msg="Compare Error"
   elif [ "${e_rs}" -ne 0 ]; then
     msg="Skip"
-  elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+  elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
     msg="Error"
   elif [ "${e_to}" -ne 0 ]; then
     msg="Warning"
@@ -3520,7 +3512,7 @@ function luks_test()
           msg="Compare Error"
         elif [ "${e_rs}" -ne 0 ]; then
           msg="Skip"
-        elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+        elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
           msg="Error"
         elif [ "${e_to}" -ne 0 ]; then
           msg="Warning"
@@ -3666,7 +3658,7 @@ function luks_legacy_test()
               msg="Compare Error"
             elif [ "${e_rs}" -ne 0 ]; then
               msg="Skip"
-            elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+            elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
               msg="Error"
             elif [ "${e_to}" -ne 0 ]; then
               msg="Warning"
@@ -3781,7 +3773,7 @@ function luks2_test()
         msg="Compare Error"
       elif [ "${e_rs}" -ne 0 ]; then
         msg="Skip"
-      elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ]; then
+      elif [ "${e_nf}" -ne 0 ] || [ "${e_nm}" -ne 0 ] || [ "${cnt}" -eq 0 ]; then
         msg="Error"
       elif [ "${e_to}" -ne 0 ]; then
         msg="Warning"
@@ -4152,7 +4144,7 @@ if [ "${PACKAGE}" -eq 0 ] || [ -z "${PACKAGE_FOLDER}" ]; then
     if [ "${HT}" -eq 65535 ]; then
       for TMP_HT in ${HASH_TYPES}; do
 
-        if ! is_in_array "${TMP_HT}" ${LUKS_MODES}; then
+        if ! ( is_in_array "${TMP_HT}" ${LUKS1_ALL_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ); then
           if ! ( is_in_array "${TMP_HT}" ${LUKS2_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ); then
             if ! is_in_array "${TMP_HT}" ${TC_MODES}; then
               if ! is_in_array "${TMP_HT}" ${VC_MODES}; then
@@ -4171,7 +4163,7 @@ if [ "${PACKAGE}" -eq 0 ] || [ -z "${PACKAGE_FOLDER}" ]; then
           continue
         fi
 
-        if ! is_in_array "${TMP_HT}" ${LUKS_MODES}; then
+        if ! ( is_in_array "${TMP_HT}" ${LUKS1_ALL_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ); then
           if ! ( is_in_array "${TMP_HT}" ${LUKS2_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ); then
             if ! is_in_array "${TMP_HT}" ${TC_MODES}; then
               if ! is_in_array "${TMP_HT}" ${VC_MODES}; then
@@ -4323,7 +4315,7 @@ if [ "${PACKAGE}" -eq 0 ] || [ -z "${PACKAGE_FOLDER}" ]; then
               truecrypt_test "${hash_type}" 0
               truecrypt_test "${hash_type}" 1
               truecrypt_test "${hash_type}" 2
-            elif is_in_array "${hash_type}" ${LUKS_MODES}; then
+            elif is_in_array "${hash_type}" ${LUKS1_ALL_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ; then
               # run luks tests
               if [ ${hash_type} -eq 14600 ]; then
                 # for legacy mode
@@ -4406,7 +4398,7 @@ if [ "${PACKAGE}" -eq 1 ]; then
     copy_cl_dir=1
   else
     for TMP_HT in $(seq "${HT_MIN}" "${HT_MAX}"); do
-      if is_in_array "${TMP_HT}" ${LUKS_MODES}; then
+      if is_in_array "${TMP_HT}" ${LUKS1_ALL_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]] ; then
         copy_luks_dir=1
       elif is_in_array "${TMP_HT}" ${LUKS2_MODES} && [[ "${GENERATE_CONTAINERS}" -eq 0 ]]; then
         copy_luks2_dir=1
